Move error log context and add event handler error
