+++ src/blocks/workbench.rs
@@ -8,8 +8,8 @@ mod options_menu;
 // mod search_menu;1
 use crate::components::lists::{ListDetails, ListElement};
 use crate::components::menu::vertical_menu::VerticalMenu;
+use crate::errors::ErrorLogContext;
 use about_menu::AboutMenu;
-use error_log::ErrorLogContext;
 use error_log::ErrorMenu;
 use export_menu::ExportMenu;
 use filter_menu::FilterMenu;
