+++ src/lib.rs
@@ -16,7 +16,7 @@ pub use wasm_bindgen_rayon::init_thread_pool;
 pub mod app;
 pub mod blocks;
 pub mod components;
-pub mod error;
+pub mod errors;
 pub mod hydration_scripts;
 pub mod pages;
